Added `channels` to services
Services have channels so that modifiers (for instance) could set a callback on a given channel, and by that a service could become more autonomic in the way of open communication with a server e.g. - via websocket. The service could handle all corresponding with the server itself and simply send a unified object via a dedicated channel.

To set a channel:
{
  channels: {
    announce(...) {
      //...
      return 'Announcements';
    }
  }
}

To broadcast through channel, only from the service scope itself:
{
  requests: {
    doStuffThatLaterOnNeedsToBeAnnounced(resolve, reject, ...) {
      this.channels.announce();
    }
  }
}

And then:
SepCon.services('...').requests.doStuffThatLaterOnNeedsToBeAnnounced();

And the use will be made like that:
SepCon.service('...').channels.announce('SOME-UNIQUE-ID', (announcement) => {
  console.log(announcement); // 'Announcements'
});

To unsubscribe:
SepCon.service('...').channels.announce('SOME-UNIQUE-ID', null);