Made caching of service requests get stored according to the definition
SepCon.createService({id:'...'}, {
  cache: {
    requests: {
      someRequestName: {
        storage: 'local' (default) // false/'local'/'storage'/'cookie',
        duration: false (default) // ms
      }
    }
  }
});

The 'someRequestName' will get arguments, according to them the response will be saved into cache. Next time these same arguments will be sent to the request - the cached result will be the one that will be sent to the .then() method, and the request itself will not even get invoked.