resume vs destroy, refProps supports arrays, multiple args for channels
Handled the resume/destroy racing conditions causing issues when re-rendering a component with referenced properties. If such a property would change while the component is not longer in DOM it would have caused inconsistency, therefore there is now a proper check, and any relevant resume/change callbacks will be stored to be executed once the component is resumed.

refProps can now get an array to easily pass props forward to nested components if the property name is the same as of the referenced one. e.g. -
prop1: 'prop1' will be equal to ['prop1'].

Channels could return an array of values to be parsed as the subscribers callback method's arguments.