Added 'endpoints', Provider's scope have 'services', enhanced 'events'...
* All "createX" returned objects (except the "createComponent"):
In addition to 'proto' and 'id' now will also add all methods defined in an object named 'endpoints'.
These methods will be executed in the actual object (which is why only Components won't have this ability - these are the only elements that are not singletons)
e.g.
definition -
{
    endpoints: {
        someMethod() {...}
    }
}

returned object -
someObject.proto // { definition: { endpoints: someMethod() } }
someObject.id // some id
someObject.someMethod() // will be executed in the original defined object's scope

* Provider:
Added references to all bind services for a given provider. Will be accessed under the property 'services'.
All nested properties and methods of each service will be accessible as if it will be in the scope of the service itself.
e.g.
this.services['some-service-id'].requests.someRequest()
or
this.services['some-service-id'].somePrivateMethod()

* Component View:
Added ability to bind events to the SepCon DOM element itself (i.e <x-sepcon-...>), by simply not specifying a selector.
e.g.
events: [
    {event: 'click', callback: 'onClick'}
]

* Component-State:
Fixed bug with objects passed to setProps and being mutated afterwards, causing the props themselves to be changed as well.