Develop (#12)
* resume vs destroy, refProps supports arrays, multiple args for channels

Handled the resume/destroy racing conditions causing issues when re-rendering a component with referenced properties. If such a property would change while the component is not longer in DOM it would have caused inconsistency, therefore there is now a proper check, and any relevant resume/change callbacks will be stored to be executed once the component is resumed.

refProps can now get an array to easily pass props forward to nested components if the property name is the same as of the referenced one. e.g. -
prop1: 'prop1' will be equal to ['prop1'].

Channels could return an array of values to be parsed as the subscribers callback method's arguments.

* bug fix - dealing with refProps arrays with nesting as refProps also

If passing by reference an array derived of a global scope (data) - and in the child component passing by reference to its own children different objects of the array - it would cause a path pointing mutation bug for the parent itself.
Fixed + added relevant tests to component-instance-state.test file

* Added 'endpoints', Provider's scope have 'services', enhanced 'events'...

* All "createX" returned objects (except the "createComponent"):
In addition to 'proto' and 'id' now will also add all methods defined in an object named 'endpoints'.
These methods will be executed in the actual object (which is why only Components won't have this ability - these are the only elements that are not singletons)
e.g.
definition -
{
    endpoints: {
        someMethod() {...}
    }
}

returned object -
someObject.proto // { definition: { endpoints: someMethod() } }
someObject.id // some id
someObject.someMethod() // will be executed in the original defined object's scope

* Provider:
Added references to all bind services for a given provider. Will be accessed under the property 'services'.
All nested properties and methods of each service will be accessible as if it will be in the scope of the service itself.
e.g.
this.services['some-service-id'].requests.someRequest()
or
this.services['some-service-id'].somePrivateMethod()

* Component View:
Added ability to bind events to the SepCon DOM element itself (i.e <x-sepcon-...>), by simply not specifying a selector.
e.g.
events: [
    {event: 'click', callback: 'onClick'}
]

* Component-State:
Fixed bug with objects passed to setProps and being mutated afterwards, causing the props themselves to be changed as well.

* 2.5.2